requirements.txt
fastapi
uvicorn
jinja2
python-multipart
SQLAlchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
httpx
pytest
pytest-asyncio
pytest-httpx
